+++ src/main/java/com/programmers/App.java
@@ -1,9 +1,12 @@
 package com.programmers;
 
-public class Application {
+import com.programmers.engine.JavaCalculator;
+import com.programmers.engine.model.Result;
+
+public class App {
     public static void main(String[] args) {
         Console console = new Console();
-        CalculateResult calculateResult = new CalculateResult();
+        Result calculateResult = new Result();
         new JavaCalculator(console, console, calculateResult).run();
     }
 }
