--- src/main/java/com/programmers/model/Result.java
@@ -1,6 +1,0 @@
-package com.programmers.model;
-
-public interface Result {
-    void save(String result);
-    void readAllResults();
-}
